lift store into client.
