+++ client/src/snapshot.rs
@@ -8,7 +8,6 @@ use serde::{Deserialize, Serialize};
 use engine::{
     snapshot,
     snapshot::{read_from, write_to, Key},
-    store::Cache,
     vault::{Key as PKey, ReadResult},
 };
 
@@ -31,7 +30,6 @@ pub struct SnapshotState(
             Client,
             HashMap<VaultId, PKey<Provider>>,
             HashMap<PKey<Provider>, Vec<ReadResult>>,
-            Cache<Vec<u8>, Vec<u8>>,
         ),
     >,
 );
@@ -49,16 +47,10 @@ impl Snapshot {
         Client,
         HashMap<VaultId, PKey<Provider>>,
         HashMap<PKey<Provider>, Vec<ReadResult>>,
-        Cache<Vec<u8>, Vec<u8>>,
     ) {
         match self.state.0.remove(&id) {
             Some(t) => t,
-            None => (
-                Client::new(id),
-                HashMap::default(),
-                HashMap::default(),
-                Cache::default(),
-            ),
+            None => (Client::new(id), HashMap::default(), HashMap::default()),
         }
     }
 
@@ -106,7 +98,6 @@ impl SnapshotState {
             Client,
             HashMap<VaultId, PKey<Provider>>,
             HashMap<PKey<Provider>, Vec<ReadResult>>,
-            Cache<Vec<u8>, Vec<u8>>,
         ),
     ) -> Self {
         let mut state = HashMap::new();
@@ -122,7 +113,6 @@ impl SnapshotState {
             Client,
             HashMap<VaultId, PKey<Provider>>,
             HashMap<PKey<Provider>, Vec<ReadResult>>,
-            Cache<Vec<u8>, Vec<u8>>,
         ),
     ) {
         self.0.insert(id, data);
